Print to Printer settings
There seems to be an issue when printing to a printer if the report has the option No Printer ticked in the page setup. For more info see comment and the SAP thread from https://stackoverflow.com/a/69991805

This update should allow the report to be printed to the specified printer even if the No Printer option is ticked.